requirements.txt
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==1.4.48
typing_extensions==4.8.0
Werkzeug==2.0.0